Move SearchForm to _pub_shared and use it in web_app.
