+++ app/lib/frontend/handlers/listing.dart
@@ -4,12 +4,11 @@
 
 import 'dart:async';
 
+import 'package:_pub_shared/search/search_form.dart';
 import 'package:shelf/shelf.dart' as shelf;
 
 import '../../package/name_tracker.dart';
 import '../../package/search_adapter.dart';
-import '../../search/search_form.dart';
-import '../../search/search_service.dart';
 import '../../shared/handlers.dart';
 import '../../shared/tags.dart';
 import '../../shared/utils.dart' show DurationTracker;
