+++ app/lib/search/search_service.dart
@@ -5,6 +5,7 @@
 import 'dart:async';
 import 'dart:math' show max;
 
+import 'package:_pub_shared/search/search_form.dart';
 import 'package:clock/clock.dart';
 import 'package:json_annotation/json_annotation.dart';
 
@@ -16,20 +17,12 @@ const int _minSearchLimit = 10;
 const int searchIndexNotReadyCode = 600;
 const String searchIndexNotReadyText = 'Not ready yet.';
 
-/// The number of packages we are going to display on a search page.
-const int resultsPerPage = 10;
-
 /// The number of page links we display, e.g. on page 10, we display direct
 /// links from page 5 to page 15.
 const int maxPageLinks = 10;
 
 /// The maximum length of the search query's text phrase that we'll try to serve.
 const _maxQueryLength = 256;
-
-/// The tag prefixes that we can detect in the user-provided search query.
-final _detectedTagPrefixes = <String>{
-  ...allowedTagPrefixes.expand((s) => [s, '-$s', '+$s']),
-};
 
 /// Statistics about the index content.
 class IndexInfo {
@@ -145,55 +138,6 @@ class ApiDocPage {
 
   Map<String, dynamic> toJson() => _$ApiDocPageToJson(this);
 }
-
-/// How search results should be ordered.
-enum SearchOrder {
-  /// Search score should be a weighted value of [text], [popularity], [points]
-  /// and [like], ordered decreasing.
-  top,
-
-  /// Search score should depend only on text match similarity, ordered
-  /// decreasing.
-  text,
-
-  /// Search order should be in decreasing last package creation time.
-  created,
-
-  /// Search order should be in decreasing last package updated time.
-  updated,
-
-  /// Search order should be in decreasing popularity score.
-  popularity,
-
-  /// Search order should be in decreasing like count.
-  like,
-
-  /// Search order should be in decreasing pub points.
-  points,
-}
-
-/// Returns null if [value] is not a recognized search order.
-SearchOrder? parseSearchOrder(String? value) {
-  if (value == null) {
-    return null;
-  }
-  for (final v in SearchOrder.values) {
-    if (v.name == value) return v;
-  }
-  return null;
-}
-
-final RegExp _whitespacesRegExp = RegExp(r'\s+');
-final RegExp _packageRegexp =
-    RegExp('package:([_a-z0-9]+)', caseSensitive: false);
-final RegExp _publisherRegexp =
-    RegExp(r'publisher:([_a-z0-9\.]+)', caseSensitive: false);
-final RegExp _refDependencyRegExp =
-    RegExp('dependency:([_a-z0-9]+)', caseSensitive: false);
-final RegExp _allDependencyRegExp =
-    RegExp(r'dependency\*:([_a-z0-9]+)', caseSensitive: false);
-final _tagRegExp =
-    RegExp(r'([\+|\-]?[a-z0-9]+:[a-z0-9\-_\.]+)', caseSensitive: false);
 
 String? _stringToNull(String? v) => (v == null || v.isEmpty) ? null : v;
 
@@ -351,226 +295,6 @@ class QueryValidity {
   bool get isRejected => rejectReason != null;
 }
 
-/// Filter conditions on tags.
-class TagsPredicate {
-  /// tag -> {true = required | false = prohibited}
-  final Map<String, bool> _values;
-
-  TagsPredicate._(this._values);
-
-  factory TagsPredicate(
-      {List<String>? requiredTags, List<String>? prohibitedTags}) {
-    final values = <String, bool>{};
-    requiredTags?.forEach((tag) => values[tag] = true);
-    prohibitedTags?.forEach((tag) => values[tag] = false);
-    return TagsPredicate._(values);
-  }
-
-  /// Pre-populates the predicate with the default tags for regular search (e.g.
-  /// typing in the search box on the landing page).
-  factory TagsPredicate.regularSearch() => TagsPredicate(
-        prohibitedTags: [
-          PackageTags.isDiscontinued,
-          PackageTags.isUnlisted,
-          PackageVersionTags.isLegacy,
-        ],
-      );
-
-  bool get isEmpty => _values.isEmpty;
-  bool get isNotEmpty => _values.isNotEmpty;
-
-  bool isRequiredTag(String tag) => _values[tag] == true;
-  bool isProhibitedTag(String tag) => _values[tag] == false;
-  bool hasTag(String tag) => _values.containsKey(tag);
-  bool anyTag(bool Function(String key) fn) => _values.keys.any(fn);
-  bool hasTagPrefix(String prefix) => anyTag((t) => t.startsWith(prefix));
-  bool hasNoTagPrefix(String prefix) => !hasTagPrefix(prefix);
-
-  /// Parses [values] passed via Uri.queryParameters
-  factory TagsPredicate.parseQueryValues(List<String>? values) {
-    final p = TagsPredicate();
-    if (values == null) {
-      return p;
-    }
-    for (var tag in values) {
-      bool required = true;
-      if (tag.startsWith('-')) {
-        tag = tag.substring(1);
-        required = false;
-      } else if (tag.startsWith('+')) {
-        tag = tag.substring(1);
-      }
-      if (required) {
-        p._values[tag] = true;
-      } else {
-        p._values[tag] = false;
-      }
-    }
-    return p;
-  }
-
-  /// Appends [other] predicate to the current set of tags, and returns a new
-  /// [TagsPredicate] instance.
-  ///
-  /// If there are conflicting tag predicates, the [other] takes precedence over
-  /// this [TagsPredicate].
-  TagsPredicate appendPredicate(TagsPredicate other) {
-    final p = TagsPredicate();
-    p._values.addAll(_values);
-    p._values.addAll(other._values);
-    return p;
-  }
-
-  /// Evaluate this predicate against the list of supplied [tags].
-  /// Returns true if the predicate matches the [tags], false otherwise.
-  bool matches(List<String> tags) {
-    for (String? tag in _values.keys) {
-      final present = tags.contains(tag);
-      final required = _values[tag]!;
-      if (required && !present) return false;
-      if (!required && present) return false;
-    }
-    return true;
-  }
-
-  /// Toggles [tag] between required and absent status.
-  TagsPredicate toggleRequired(String tag) {
-    final current = _values[tag];
-    return _change(tag, current == null ? true : null);
-  }
-
-  TagsPredicate _change(String tag, bool? value) {
-    final current = _values[tag];
-    if (current == value) {
-      return this;
-    }
-    final newValues = Map<String, bool>.from(_values);
-    if (value == null) {
-      newValues.remove(tag);
-    } else {
-      newValues[tag] = value;
-    }
-    return TagsPredicate._(newValues);
-  }
-
-  /// Returns the list of tag values that can be passed to search service URL.
-  List<String> toQueryParameters() {
-    return _values.entries.map((e) => e.value ? e.key : '-${e.key}').toList();
-  }
-}
-
-class ParsedQueryText {
-  final String? text;
-  final String? packagePrefix;
-
-  /// Dependency match for direct or dev dependency.
-  final List<String> refDependencies;
-
-  /// Dependency match for all dependencies, including transitive ones.
-  final List<String> allDependencies;
-
-  /// Match the publisher of the package.
-  final String? publisher;
-
-  /// Detected tags in the user-provided query.
-  TagsPredicate tagsPredicate;
-
-  ParsedQueryText._(
-    this.text,
-    this.packagePrefix,
-    this.refDependencies,
-    this.allDependencies,
-    this.publisher,
-    this.tagsPredicate,
-  );
-
-  factory ParsedQueryText.parse(String? q) {
-    String? queryText = q ?? '';
-    queryText = ' $queryText ';
-    String? packagePrefix;
-    final Match? pkgMatch = _packageRegexp.firstMatch(queryText);
-    if (pkgMatch != null) {
-      packagePrefix = pkgMatch.group(1);
-      queryText = queryText.replaceFirst(_packageRegexp, ' ');
-    }
-
-    List<String> extractRegExp(RegExp regExp, {bool Function(String?)? where}) {
-      final values = regExp
-          .allMatches(queryText!)
-          .map((Match m) => m.group(1))
-          .where((s) => where == null || where(s))
-          .cast<String>()
-          .toList();
-      if (values.isNotEmpty) {
-        queryText = queryText!.replaceAll(regExp, ' ');
-      }
-      return values;
-    }
-
-    final List<String> dependencies = extractRegExp(_refDependencyRegExp);
-    final List<String> allDependencies = extractRegExp(_allDependencyRegExp);
-    final allPublishers = extractRegExp(_publisherRegexp);
-    final publisher = allPublishers.isEmpty ? null : allPublishers.first;
-
-    final tagValues = extractRegExp(
-      _tagRegExp,
-      where: (tag) => _detectedTagPrefixes.any((p) => tag!.startsWith(p)),
-    );
-    final tagsPredicate = TagsPredicate.parseQueryValues(tagValues);
-
-    queryText = queryText!.replaceAll(_whitespacesRegExp, ' ').trim();
-    if (queryText!.isEmpty) {
-      queryText = null;
-    }
-
-    return ParsedQueryText._(
-      queryText,
-      packagePrefix,
-      dependencies,
-      allDependencies,
-      publisher,
-      tagsPredicate,
-    );
-  }
-
-  ParsedQueryText change({
-    TagsPredicate? tagsPredicate,
-  }) {
-    return ParsedQueryText._(
-      text,
-      packagePrefix,
-      refDependencies,
-      allDependencies,
-      publisher,
-      tagsPredicate ?? this.tagsPredicate,
-    );
-  }
-
-  bool get hasAnyDependency =>
-      refDependencies.isNotEmpty || allDependencies.isNotEmpty;
-
-  bool get hasOnlyFreeText =>
-      text != null &&
-      text!.isNotEmpty &&
-      packagePrefix == null &&
-      !hasAnyDependency &&
-      publisher == null &&
-      tagsPredicate.isEmpty;
-
-  @override
-  String toString() {
-    if (hasOnlyFreeText) return text!;
-    return <String>[
-      if (packagePrefix != null) 'package:$packagePrefix',
-      ...refDependencies.map((d) => 'dependency:$d'),
-      ...allDependencies.map((d) => 'dependency*:$d'),
-      if (publisher != null) 'publisher:$publisher',
-      ...tagsPredicate.toQueryParameters(),
-      if (text != null && text!.isNotEmpty) text!,
-    ].join(' ');
-  }
-}
-
 @JsonSerializable(includeIfNull: false)
 class PackageSearchResult {
   final DateTime? timestamp;
@@ -693,17 +417,43 @@ class ApiPageRef {
   Map<String, dynamic> toJson() => _$ApiPageRefToJson(this);
 }
 
-/// Extracts the 'page' query parameter from requested URL's [queryParameters].
-///
-/// Returns a valid positive integer.
-int extractPageFromUrlParameters(Map<String, String> queryParameters) {
-  final pageAsString = queryParameters['page'];
-  final pageAsInt = int.tryParse(pageAsString ?? '1') ?? 1;
-  return max(pageAsInt, 1);
-}
-
 abstract class DependencyTypes {
   static const dev = 'dev';
   static const direct = 'direct';
   static const transitive = 'transitive';
+}
+
+extension SearchFormExt on SearchForm {
+  ServiceSearchQuery toServiceQuery() {
+    final prohibitLegacy = !context.includeAll &&
+        !parsedQuery.tagsPredicate.anyTag((tag) =>
+            tag == PackageVersionTags.isLegacy ||
+            tag == PackageVersionTags.showLegacy ||
+            tag == PackageTags.showHidden);
+    final prohibitDiscontinued = !context.includeAll &&
+        !parsedQuery.tagsPredicate.anyTag((tag) =>
+            tag == PackageTags.isDiscontinued ||
+            tag == PackageTags.showDiscontinued ||
+            tag == PackageTags.showHidden);
+    final prohibitUnlisted = !context.includeAll &&
+        !parsedQuery.tagsPredicate.anyTag((tag) =>
+            tag == PackageTags.isUnlisted ||
+            tag == PackageTags.showUnlisted ||
+            tag == PackageTags.showHidden);
+    final tagsPredicate = TagsPredicate(
+      prohibitedTags: [
+        if (prohibitDiscontinued) PackageTags.isDiscontinued,
+        if (prohibitUnlisted) PackageTags.isUnlisted,
+        if (prohibitLegacy) PackageVersionTags.isLegacy,
+      ],
+    );
+    return ServiceSearchQuery.parse(
+      query: query,
+      tagsPredicate: tagsPredicate,
+      publisherId: context.publisherId,
+      offset: offset,
+      limit: pageSize,
+      order: order,
+    );
+  }
 }
