+++ app/test/search/mem_index_test.dart
@@ -4,9 +4,9 @@
 
 import 'dart:convert';
 
+import 'package:_pub_shared/search/search_form.dart';
 import 'package:clock/clock.dart';
 import 'package:pub_dev/search/mem_index.dart';
-import 'package:pub_dev/search/search_form.dart';
 import 'package:pub_dev/search/search_service.dart';
 import 'package:test/test.dart';
